move convert_non_serializable to utils.py
