+++ leap/utils.py
@@ -16,9 +16,7 @@ LEAP_PATH = pathlib.Path(__file__).parents[1].absolute()
 
 
 class UUID4:
-    def __init__(
-        self, uuid4: uuid.UUID | None = None, short: str | None = None
-    ):
+    def __init__(self, uuid4: uuid.UUID | None = None, short: str | None = None):
         if uuid4 is None:
             uuid4 = uuid.uuid4()
         if short is None:
@@ -62,9 +60,7 @@ def logit(x: float | np.ndarray) -> float | np.ndarray:
 
 
 def compute_ordinal_regression(
-    η: float,
-    θ: float | list[float],
-    prob_function: Callable = logistic.cdf
+    η: float, θ: float | list[float], prob_function: Callable = logistic.cdf
 ) -> list[float]:
     """Compute the probability that ``y = k`` for each value of ``k``.
 
@@ -100,9 +96,9 @@ def compute_ordinal_regression(
     """
 
     if isinstance(θ, float):
-        θ = [-10**5, θ, 10**5]
+        θ = [-(10**5), θ, 10**5]
     else:
-        θ = [-10**5] + θ + [10**5]
+        θ = [-(10**5)] + θ + [10**5]
 
     return [prob_function(θ[k + 1] - η) - prob_function(θ[k] - η) for k in range(len(θ) - 1)]
 
@@ -127,15 +123,12 @@ def get_data_path(data_path: str) -> pathlib.Path:
         data_path = str(pathlib.Path(data_path).relative_to("original_data"))
         data_folder = "leap.original_data"
 
-
     package_path = str(pkg_resources.files(data_folder).joinpath(data_path))
     
     if os.path.isfile(package_path) or os.path.isdir(package_path):
         return pathlib.Path(package_path)
     else:
         raise FileNotFoundError(f"Path {data_path} not found.")
-        
-    
 
 
 def check_file(file_path: str | pathlib.Path, ext: str):
@@ -192,49 +185,76 @@ def check_year(year: int, df: pd.DataFrame):
 
 def check_province(province: str):
     """Check if the province is valid.
-    
+
     Args:
         province: The province abbreviation.
-        
+
     Raises:
         ValueError: If the province is not valid.
     """
 
-    PROVINCES = [
-        "CA", "AB", "BC", "MB", "NB", "NL", "NS", "NT", "NU", "ON", "PE", "QC", "SK", "YT"
-    ]
+    PROVINCES = ["CA", "AB", "BC", "MB", "NB", "NL", "NS", "NT", "NU", "ON", "PE", "QC", "SK", "YT"]
     if province not in PROVINCES:
         raise ValueError(f"province must be one of {PROVINCES}, received {province}")
-    
+
 
 def check_projection_scenario(projection_scenario: str):
     """Check if the projection scenario is valid.
-    
+
     Args:
         projection_scenario: The projection scenario abbreviation.
-        
+
     Raises:
         ValueError: If the projection scenario is not valid.
     """
 
-    PROJECTION_SCENARIOS = [
-        "past", "LG", "HG", "M1", "M2", "M3", "M4", "M5", "M6", "FA", "SA"
-    ]
+    PROJECTION_SCENARIOS = ["past", "LG", "HG", "M1", "M2", "M3", "M4", "M5", "M6", "FA", "SA"]
     if projection_scenario not in PROJECTION_SCENARIOS:
         raise ValueError(
             f"projection_scenario must be one of {PROJECTION_SCENARIOS}, "
             f"received {projection_scenario}"
         )
 
+
+def convert_non_serializable(obj: np.ndarray | object) -> list | str:
+    """
+    Description
+    ---
+    Convert non-serializable objects into JSON-friendly formats.
+
+    This function is intended to be used with ``json.dumps(..., default=convert_non_serializable)``.
+    It handles cases where objects cannot be directly serialized into JSON:
+
+    - NumPy arrays (``numpy.ndarray``) are converted to Python lists.
+    - Other unsupported types are converted to their string representation.
+
+    Args
+    ---
+        obj (``np.ndarray | object``): The object to be converted.
+
+    Returns
+    ---
+        ``list | str``: A JSON-serializable equivalent of ``obj``:
+            - A list if ``obj`` is an ndarray.
+            - A string representation otherwise.
+    """
+
+    if isinstance(obj, np.ndarray):
+        return obj.tolist()
+    # Default for unsupported types
+    return str(obj)
+
+
 class Sex:
     """A class to handle different formats of the ``sex`` variable."""
+
     def __init__(self, value: str | int | bool):
         """Initialize the ``Sex`` class.
-        
+
         Args:
             value: The value of the sex variable. Must be one of:
                 ``"M"``, ``"F"``, ``1``, ``0``, ``True``, or ``False``.
-        
+
         Examples:
 
             >>> sex = Sex("F")
@@ -286,20 +306,20 @@ class Sex:
                 value_bool = False
         else:
             raise TypeError(f"sex must be str, int, or bool, received {type(value)}")
-        
+
         self._value_str = value_str
         self._value_int = value_int
         self._value_bool = value_bool
 
     def __str__(self) -> str:
         return self._value_str
-    
+
     def __int__(self) -> int:
         return self._value_int
-    
+
     def __bool__(self) -> bool:
         return self._value_bool
-    
+
     def __eq__(self, value: str | int | bool) -> bool:
         if isinstance(value, str):
             return self._value_str == value
